+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# TensionDev.NMEA0183
+# TensionDev.Maritime.NMEA0183
 
 # Changelog
 All notable changes to this project will be documented in this file.
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- Changed Namespace from TensionDev.NMEA0183 to TensionDev.Maritime.NMEA0183 to reflect possible expansion to other maritime implementations.
+
 
 ## [v0.1.0-alpha] - 2021-09-10
 [v0.1.0-alpha](https://github.com/TensionDev/NMEA0183/releases/tag/v0.1.0-alpha)